LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCDtlsTransport-state-expected.txt


PASS DTLS transport goes to connected state
FAIL close() causes the local transport to close immediately assert_equals: expected "closed" but got "connected"
PASS close() causes the other end's DTLS transport to close
FAIL stop bundled transceiver retains dtls transport state assert_equals: getReceivers does not expose a receiver of a stopped transceiver expected 1 but got 2
